sample-fabric-streaming/emulator/requirement.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
configargparse